requirements.txt
pillow
ConfigArgParse